Refactor: moving ColorEditor class to pro.sketchware
