+++ app/src/main/AndroidManifest.xml
@@ -300,7 +300,7 @@
             android:configChanges="orientation|screenSize" />
 
         <activity android:name="mod.trindadedev.ui.activities.SettingsActivity" />
-        <activity android:name="mod.mustiy.ColorEditorActivity" />
+        <activity android:name="pro.sketchware.activities.color.editor.ColorEditorActivity" />
         <activity android:name="mod.bobur.StringEditorActivity"
             android:windowSoftInputMode="stateHidden|adjustPan"/>
